+++ Employees/_template.txt
@@ -1,0 +1,47 @@
+351
+492
+500
+43
+22 137 2.465957796526439
+5 133 5.255238207666245
+14 172 5.839629131863049
+14 101 3.3946350110888446
+9 175 2.38733551568032
+25 239 2.8636300578830216
+1 226 4.051596007594391
+5 242 5.808008433161021
+14 256 5.198262348928523
+1 256 4.421738481170593
+10 254 2.7010417169844985
+16 223 1.054445996961583
+22 223 0.2680787823752766
+21 250 3.496107358076138
+29 248 3.2685942071490857
+28 253 0.3862870404548815
+43 241 1.6366941189399213
+47 245 5.667237335333198
+59 225 3.801657639090294
+62 234 2.2215021799854213
+73 217 0.629621994207433
+71 245 3.897138341167903
+84 215 0.9550019328696848
+76 192 0.22030573022441047
+58 207 5.637993720509968
+85 163 2.9966049469699954
+60 169 1.6845579444498924
+43 169 3.173216661067414
+45 154 5.462457300806854
+54 256 1.5660086719569932
+42 256 1.4486035795781689
+14 256 3.6779183182244495
+15 256 4.880475393168041
+10 256 4.423491044723953
+6 256 3.302447978205045
+4 256 1.2570061480465924
+27 256 3.436042069748659
+33 256 3.7513009119681726
+39 256 1.582930600517244
+13 256 0.9878366801443832
+52 256 6.184970375468332
+63 256 0.4150801237391306
+48 256 1.0924993726544088
