+++ main.cs
@@ -4,55 +4,45 @@ using System.Linq;
 using System.Globalization;
 using BioLab.Biometrics.Mcc.Sdk;
 
-namespace BioLab.Biometrics.Mcc.Sdk;
-class Program
+const string Employees = "Employees";
+templateCreation tp = new templateCreation();
+Matcher mr = new Matcher();
+string fingerInput = @"SampleMinutiae\1_1.txt";
+bool running = true;
+mr.loadParams();
+while (running)
 {
+    Console.WriteLine("Enter ID:");
+    string employeeId = Console.ReadLine();
+    string employeeDirectory = Path.Combine(Employees, employeeId);
+    
 
-    static void Main()
+    if (Directory.Exists(employeeDirectory))
     {
-        const string Employees = "Employees"; //try enroll first explicit
-        MccSdk.SetMccEnrollParameters(@"Biolab\Mcc\Sdk\MccPaperEnrollParameters.xml");
-        MccSdk.SetMccMatchParameters(@"Biolab\Mcc\Sdk\MccPaperMatchParameters.xml");
-        var template1 = MccSdk.CreateMccTemplateFromTextTemplate(@"SampleMinutiae\1_1.txt");
-        templateCreation tp = new templateCreation();
-        bool running = true;
-        while (running)
+        var templateFiles = Directory.GetFiles(employeeDirectory, "*.txt");
+        if (templateFiles.Length > 0)
         {
-            Console.WriteLine("Enter ID:");
-            string employeeId = Console.ReadLine();
-            string employeeDirectory = Path.Combine(Employees, employeeId);
-
-            if (Directory.Exists(employeeDirectory))
-            {
-                string[] templateFiles = Directory.GetFiles(employeeDirectory, "*.txt");
-                if (templateFiles.Length > 0)
-                {   
-                    var template2 = MccSdk.CreateMccTemplateFromTextTemplate(@"Employees\1\1_template.txt"); 
-                    Console.WriteLine("Here we run the matcher to see if the templates match");
-                    var score = MccSdk.MatchMccTemplates(template1, template2);
-                    Console.WriteLine(string.Format(CultureInfo.InvariantCulture,
-                    "Match score: {0:R}",
-                    score));
+            
+            Console.WriteLine("Here we run the matcher to see if the templates match");
+            mr.matchTemplates(fingerInput, templateFiles.ToString());
+            Console.WriteLine(templateFiles);
 
 
-                }
-                else
-                {
-                    Console.WriteLine("No template found, creating a new fingerprint template...");
-                    // Create the template and get the combined matrix
-                    double[][] combinedMatrix = tp.CreateTemplate(employeeId);
-
-                    // Save the combined matrix to a file
-                    tp.SaveTemplateToFile(employeeId, combinedMatrix);
-
-                    //TODO: Implement matching from MccSDK
-                }
-            }
-            else
-            {
-                Console.WriteLine("ID does not match employee");
-                Console.WriteLine("Make sure correct ID is entered or contact administrator");
-            }
         }
+        else
+        {
+            Console.WriteLine("No template found, creating a new fingerprint template...");
+            // Create the template and get the combined matrix
+            double[][] combinedMatrix = tp.CreateTemplate(employeeId);
+
+            // Save the combined matrix to a file
+            tp.SaveTemplateToFile(employeeId, combinedMatrix);
+
+        }
+    }
+    else
+    {
+        Console.WriteLine("ID does not match employee");
+        Console.WriteLine("Make sure correct ID is entered or contact administrator");
     }
 }
